rename - path::detail to path::parse and change the file structure
